requirements.txt
numpy 
tensorflow
torch
opencv-python
transformers
matplotlib
scikit-learn
torchvision
pandas
seaborn
Pillow
flask